rename Book components to ClickableBooks
